pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "audio-visualizer-pro"
version = "2.0.0"
description = "Modulares, KI-optimiertes Audio-Visualisierungs-System fuer professionelle Musikvideos und Podcast-Visuals"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Audio Visualizer Pro Team"}
]
keywords = [
    "audio", "visualization", "music", "video", "gpu", "opengl", 
    "ffmpeg", "podcast", "music-video"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Content Creators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "numpy>=1.21.0",
    "Pillow>=9.0.0",
    "librosa>=0.10.0",
    "pydantic>=1.10.0",
    "click>=8.0.0",
    "moderngl>=5.10.0",
    "glcontext>=2.5.0",
    "scipy>=1.7.0",
    "google-genai>=1.0.0",
    "python-dotenv>=0.19.0",
    "streamlit>=1.28.0",
    "pytest>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[project.scripts]
audio-visualizer = "main:cli"
avp = "main:cli"

[project.urls]
Homepage = "https://github.com/audio-visualizer-pro/audio-visualizer-pro"
Documentation = "https://github.com/audio-visualizer-pro/audio-visualizer-pro/blob/main/README.md"
Repository = "https://github.com/audio-visualizer-pro/audio-visualizer-pro"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
